Create: 将几何和驱动问题对象分成geometry和generate两个py文件
